Chapter10/code/requirements.txt
google-cloud-speech
google-translate
pandas